requirements.txt
python-dotenv
asyncio
telebot
aiohttp